Fix scrolling when inline-editing a news
Go to the appropriate anchor when toggling the inline news edition form
for the news to always be visible, even if the news and form have very
different sizes.

This was not a problem before since the edit link was always above the
news and thus would always stay visible, but now it is shown below, the
form may not be in the screen after showing it.