Index: flush page top if we're starting downloading commits feed
This prevents the user from waiting too long to see anything simply
because we're downloading a feed that might take long to come.

So at least the user will see the part of the page we output until
now (heading, screenshot, last news, ...), waiting only for bottom
part and making the overall impression ways less laggy.

Just note this changes only applies when the cache for the commits
feed expired, which is currently once an hour.