Switch to a raw PHP template system
PHPTemplate has an interface similar to FileTemplate but simply wraps
PHP inclusion and making a defined variable set available to the
template.

The advantages over the current template solution are that it's faster
(it's simply PHP, no additional layer) and more powerful (the whole PHP
syntax and functions are available).  On the other hand, the syntax is
a bit less compact (using "<?php ... ?>" rather than "{...}").