Change tag delimiter from ' ' (space) to ',' (comma)
This changes the GET API of the medias page, but it wasn't official
nor used anywhere so it shuold be harmless.

It also requires an update of the medias database table ('bse_medias'
by default) to use that new separator.  This update must be done
manually with something like the following query:

	UPDATE `bse_medias` SET `tags`=REPLACE(`tags`, ' ', ',');