Medias: fix media array escaping not to create values
media_escape_db_array() used to sanitize all fields it knows they
should be sanitized regardless they were present or not in the array.

This brakes the media updating code because the ID isn't updated and
thus isn't in the array, which leads to the creation of the element
with a default (trash) value that further code may use.

Fixed it so it only escapes elements that are actually present, which
then fixes the problem.