Update MyDB to use mysqli
Rewrite the MyDB class to use the object-oriented mysqli API rather
than the old procedural mysql API.  Now MyDB is simply a layer on top
of mysqli rather than completely separate class.

The API is almost a drop-in replacement of the old MyDB class with only
a few small updates or unused features removed.

The calling code was updated where appropriate and cleaned up a bit.