Pause the homepage slideshow when the page doesn't have focus
This fixes some rendering issues when the pages regains the focus under
Mozilla Firefox, because that browser clamps timeouts when they run in
background and runs all missed ones when the page regains focus.

This also makes the slideshow use less resources on the user computer
if the page is not visible.